# server/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MERN-stack Swagger",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://fullstack-reactapp.herokuapp.com/api/v1"
    },
    {
      "url": "http://localhost:8000/api/v1"
    }
  ],
...
  "paths": {
...
    "/auth/checkToken": {
      "get": {
        "tags": ["Authorization"],
        "summary": "Check if token is valid",
        "operationId": "checkToken",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "authenticated": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "object"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "Basic JWT authorization.",
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}